+++ medias/medias.lua
@@ -9,3 +9,7 @@ C["Medias"].ArrowDown = [[Interface\AddOns\Tukui\Medias\Textures\Others\ArrowDow
 C["Medias"].ArrowUp = [[Interface\AddOns\Tukui\Medias\Textures\Others\ArrowUp]]
 C["Medias"].Mail = [[Interface\AddOns\Tukui\Medias\Textures\Others\Mail]]
 C["Medias"].RaidIcons = [[Interface\AddOns\Tukui\Medias\Textures\Others\RaidIcons]]
+
+C["Medias"].roleDPS = [[Interface\AddOns\LuaUI\medias\textures\roleDPS.tga]]
+C["Medias"].roleTANK = [[Interface\AddOns\LuaUI\medias\textures\roleTANK.tga]]
+C["Medias"].roleHEALER = [[Interface\AddOns\LuaUI\medias\textures\roleHEALER.tga]]